feat: show budget details in modal
